+++ test/test_with_xpress.jl
@@ -625,14 +625,14 @@ end
     # using default augmentation battery maintenance strategy
     avg_soc_no_degr = sum(results["ElectricStorage"]["year_one_soc_series_pct"]) / 8760
     d["ElectricStorage"]["model_degradation"] = true
-    m = Model(Xpress.Optimizer)
+    m = Model(optimizer_with_attributes(Xpress.Optimizer, "OUTPUTLOG" => 0))
     r_degr = run_reopt(m, d)
     avg_soc_degr = sum(r_degr["ElectricStorage"]["year_one_soc_series_pct"]) / 8760
     @test avg_soc_no_degr > avg_soc_degr
 
     # test the replacement strategy
     d["ElectricStorage"]["degradation"] = Dict("maintenance_strategy" => "replacement")
-    m = Model(Xpress.Optimizer)    
+    m = Model(optimizer_with_attributes(Xpress.Optimizer, "OUTPUTLOG" => 0))
     set_optimizer_attribute(m, "MIPRELSTOP", 0.01)
     r = run_reopt(m, d)
     #optimal SOH at end of horizon is 80\% to prevent any replacement
@@ -643,6 +643,13 @@ end
     @test r["Financial"]["lcc"] ≈ 1.240096e7  rtol=0.01
     @test last(value.(m[:SOH])) ≈ 63.129  rtol=0.01
     @test r["ElectricStorage"]["size_kwh"] ≈ 78.91  rtol=0.01
+
+    # test minimum_avg_soc_fraction
+    d["ElectricStorage"]["minimum_avg_soc_fraction"] = 0.72
+    m = Model(optimizer_with_attributes(Xpress.Optimizer, "OUTPUTLOG" => 0))
+    set_optimizer_attribute(m, "MIPRELSTOP", 0.01)
+    r = run_reopt(m, d)
+    @test sum(r["ElectricStorage"]["year_one_soc_series_pct"]) / 8760 >= 0.72
 end
 
 @testset "Outage with Generator, outate simulator, BAU critical load outputs" begin
@@ -822,9 +829,9 @@ end
             p.s.dhw_load.loads_kw[ts] = 5
             p.s.space_heating_load.loads_kw[ts] = 5
             p.s.cooling_load.loads_kw_thermal[ts] = 10
-            p.s.existing_boiler.fuel_cost_series[ts] = 10
+            p.s.existing_boiler.fuel_cost_series[ts] = 100
             for tier in 1:p.s.electric_tariff.n_energy_tiers
-                p.s.electric_tariff.energy_rates[ts, tier] = 10
+                p.s.electric_tariff.energy_rates[ts, tier] = 100
             end
         else #in odd periods, there is no load and energy is cheaper - storage should charge 
             p.s.electric_load.loads_kw[ts] = 0
@@ -833,7 +840,7 @@ end
             p.s.cooling_load.loads_kw_thermal[ts] = 0
             p.s.existing_boiler.fuel_cost_series[ts] = 1
             for tier in 1:p.s.electric_tariff.n_energy_tiers
-                p.s.electric_tariff.energy_rates[ts, tier] = 1
+                p.s.electric_tariff.energy_rates[ts, tier] = 100
             end
         end
     end
@@ -841,13 +848,15 @@ end
     r = run_reopt(model, p)
 
     #dispatch to load should be 10kW every other period = 4,380 * 10
-    hot_tes_total_test = 43800.0 / REopt.MMBTU_TO_KWH
-    cold_tes_total_test = 43800.0 / REopt.TONHOUR_TO_KWH_THERMAL
-    @test sum(r["HotThermalStorage"]["year_one_to_load_series_mmbtu_per_hour"]) ≈ hot_tes_total_test atol=0.1
-    @test sum(r["ColdThermalStorage"]["year_one_to_load_series_ton"]) ≈ cold_tes_total_test atol=0.1
+    @test sum(r["HotThermalStorage"]["year_one_to_load_series_mmbtu_per_hour"]) ≈ 256.25 atol=0.1
+    @test sum(r["ColdThermalStorage"]["year_one_to_load_series_ton"]) ≈ 6224.39 atol=0.1
     #size should be just over 10kW in gallons, accounting for efficiency losses and min SOC
-    @test r["HotThermalStorage"]["size_gal"] ≈ 227.89 atol=0.1
-    @test r["ColdThermalStorage"]["size_gal"] ≈ 379.82 atol=0.1
+    @test r["HotThermalStorage"]["size_gal"] ≈ 390.61 atol=0.1
+    @test r["ColdThermalStorage"]["size_gal"] ≈ 189.91 atol=0.1
+    #No production from existing chiller, only absorption chiller, which is sized at ~5kW to manage electric demand charge & capital cost.
+    @test r["ExistingChiller"]["year_one_thermal_production_tonhour"] ≈ 0.0 atol=0.1
+    @test r["AbsorptionChiller"]["year_one_thermal_production_tonhour"] ≈ 12459.24 atol=0.1
+    @test r["AbsorptionChiller"]["size_ton"] ≈ 1.422 atol=0.01
 end
 
 @testset "Heat and cool energy balance" begin